requirements.txt
blinker
python_dotenv
watchdog
flask
flask_restful
requests
psycopg2-binary
